all: Add ISM_2400 band definition
