+++ pkg/band/ism_2400.go
@@ -1,4 +1,4 @@
-// Copyright © 2019 The Things Network Foundation, The Things Industries B.V.
+// Copyright © 2021 The Things Network Foundation, The Things Industries B.V.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,21 +14,17 @@
 
 package band
 
-import (
-	"go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
+import "go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
+
+const (
+	// ISM_2400 is the ID of the LoRa 2.4 GHz band.
+	ISM_2400 = "ISM_2400"
+
+	ism2400BeaconFrequency = 2424000000
 )
 
-//revive:disable:var-naming
-
-var ism_2400 Band
-
-// ISM_2400 is the ID of the LoRa 2.4 GHz band.
-const ISM_2400 = "ISM_2400"
-
-//revive:enable:var-naming
-
-func init() {
-	defaultChannels := []Channel{
+var (
+	ism2400DefaultChannels = []Channel{
 		{
 			Frequency:   2403000000,
 			MaxDataRate: ttnpb.DATA_RATE_7,
@@ -42,9 +38,8 @@ func init() {
 			MaxDataRate: ttnpb.DATA_RATE_7,
 		},
 	}
-	const beaconFrequency = 2424000000
 
-	downlinkDRTable := [8][6]ttnpb.DataRateIndex{
+	ism2400DownlinkDRTable = [8][6]ttnpb.DataRateIndex{
 		{0, 0, 0, 0, 0, 0},
 		{1, 0, 0, 0, 0, 0},
 		{2, 1, 0, 0, 0, 0},
@@ -54,93 +49,4 @@ func init() {
 		{6, 5, 4, 3, 2, 1},
 		{7, 6, 5, 4, 3, 2},
 	}
-
-	ism_2400 = Band{
-		ID: ISM_2400,
-
-		MaxUplinkChannels: 16,
-		UplinkChannels:    defaultChannels,
-
-		MaxDownlinkChannels: 16,
-		DownlinkChannels:    defaultChannels,
-
-		SubBands: []SubBandParameters{
-			{
-				MinFrequency: 2400000000,
-				MaxFrequency: 2500000000,
-				DutyCycle:    1,
-				MaxEIRP:      8.0 + eirpDelta,
-			},
-		},
-
-		DataRates: map[ttnpb.DataRateIndex]DataRate{
-			ttnpb.DATA_RATE_0: makeLoRaDataRate(12, 812000, makeConstMaxMACPayloadSizeFunc(59)),
-			ttnpb.DATA_RATE_1: makeLoRaDataRate(11, 812000, makeConstMaxMACPayloadSizeFunc(123)),
-			ttnpb.DATA_RATE_2: makeLoRaDataRate(10, 812000, makeConstMaxMACPayloadSizeFunc(230)),
-			ttnpb.DATA_RATE_3: makeLoRaDataRate(9, 812000, makeConstMaxMACPayloadSizeFunc(230)),
-			ttnpb.DATA_RATE_4: makeLoRaDataRate(8, 812000, makeConstMaxMACPayloadSizeFunc(230)),
-			ttnpb.DATA_RATE_5: makeLoRaDataRate(7, 812000, makeConstMaxMACPayloadSizeFunc(230)),
-			ttnpb.DATA_RATE_6: makeLoRaDataRate(6, 812000, makeConstMaxMACPayloadSizeFunc(230)),
-			ttnpb.DATA_RATE_7: makeLoRaDataRate(5, 812000, makeConstMaxMACPayloadSizeFunc(230)),
-		},
-
-		ReceiveDelay1:        defaultReceiveDelay1,
-		ReceiveDelay2:        defaultReceiveDelay2,
-		JoinAcceptDelay1:     defaultJoinAcceptDelay1,
-		JoinAcceptDelay2:     defaultJoinAcceptDelay2,
-		MaxFCntGap:           defaultMaxFCntGap,
-		ADRAckLimit:          defaultADRAckLimit,
-		ADRAckDelay:          defaultADRAckDelay,
-		MinRetransmitTimeout: defaultRetransmitTimeout - defaultRetransmitTimeoutMargin,
-		MaxRetransmitTimeout: defaultRetransmitTimeout + defaultRetransmitTimeoutMargin,
-
-		DefaultMaxEIRP: 10,
-		TxOffset: []float32{
-			0,
-			-2,
-			-4,
-			-6,
-			-8,
-			-10,
-			-12,
-			-14,
-		},
-
-		Rx1Channel: channelIndexIdentity,
-		Rx1DataRate: func(idx ttnpb.DataRateIndex, offset ttnpb.DataRateOffset, _ bool) (ttnpb.DataRateIndex, error) {
-			if idx > ttnpb.DATA_RATE_7 {
-				return 0, errDataRateIndexTooHigh.WithAttributes("max", 7)
-			}
-			if offset > 5 {
-				return 0, errDataRateOffsetTooHigh.WithAttributes("max", 5)
-			}
-			return downlinkDRTable[idx][offset], nil
-		},
-
-		GenerateChMasks: generateChMask16,
-		ParseChMask:     parseChMask16,
-
-		LoRaCodingRate: "4/8LI",
-
-		FreqMultiplier:   200,
-		ImplementsCFList: true,
-		CFListType:       ttnpb.CFListType_FREQUENCIES,
-
-		DefaultRx2Parameters: Rx2Parameters{ttnpb.DATA_RATE_0, 2423000000},
-
-		Beacon: Beacon{
-			DataRateIndex:    ttnpb.DATA_RATE_3,
-			CodingRate:       "4/8LI",
-			ComputeFrequency: func(_ float64) uint64 { return beaconFrequency },
-		},
-		PingSlotFrequency: uint64Ptr(beaconFrequency),
-
-		regionalParameters1_v1_0:       bandIdentity,
-		regionalParameters1_v1_0_1:     bandIdentity,
-		regionalParameters1_v1_0_2:     bandIdentity,
-		regionalParameters1_v1_0_2RevB: bandIdentity,
-		regionalParameters1_v1_0_3RevA: bandIdentity,
-		regionalParameters1_v1_1RevA:   bandIdentity,
-	}
-	All[ISM_2400] = ism_2400
-}
+)
